requirements-mem0.txt
# =============================================================================
# Mem0 Dependencies for AGI Pipeline v3
# =============================================================================
# Install with: pip install -r requirements-mem0.txt
# Or in conda: pip install -r requirements-mem0.txt --break-system-packages
# =============================================================================

# Core Mem0 package (includes qdrant-client)
mem0ai>=1.0.0

# YAML config loading
pyyaml>=6.0

# These are typically included with mem0ai, but explicit for clarity:
qdrant-client>=1.7.0
pydantic>=2.0.0

# Optional: For async operations
aiohttp>=3.8.0
